+++ docs/usage/security/etcd_encryption_config.md
@@ -19,7 +19,7 @@ The `spec.kubernetes.kubeAPIServer.encryptionConfig` field in the Shoot API allo
     - `aesgcm`
     - `secretbox`
   - The default encryption provider is `aescbc`.
-  - Only newly created `Shoots` can specify the `aesgcm` and `secretbox` encryption providers.
+  - Changing the encryption provider type triggers an encryption key rotation.
   - **Important for `aesgcm`**: The `aesgcm` provider uses 96-bit IVs (nonces), and per [NIST SP 800-38D](https://csrc.nist.gov/pubs/sp/800/38/d/final), the total number of encryption invocations with a given key should not exceed 2³². To mitigate the risk of nonce collisions, Gardener defaults and requires ETCD encryption key auto-rotation to be enabled. By default, the rotation period is set to **28 days** for newly created Shoots. The maximum allowed rotation period for `aesgcm` is **90 days**.
 
 ## Example Usage in a `Shoot`
